% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter.R
\name{apply_filter}
\alias{apply_filter}
\title{apply_filter}
\usage{
apply_filter(image_file, filter_func, cutoff = 127)
}
\description{
This function applies a filter of choice, either a grayscale filter, or cutoff filter that color everything above cut off pink. Cutoff defaults to 127 if not specified.
}
\examples{
Usage of this function
  apply_filter("samples/Test image-1.jpg","grayscale")
  apply_filter("samples/Test image-1.jpg","cutoff") #defaults to cutoff 127
  apply_filter("samples/Test image-1.jpg","cutoff",20)

}
