% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgrayscale.R
\name{Rgrayscale}
\alias{Rgrayscale}
\title{Rgrayscale}
\usage{
Rgrayscale(imageFile)
}
\arguments{
\item{imageFile}{path to image to be converted to grayscale}
}
\value{
returns a grayscale image
}
\description{
this function should use Rloadimage to load image and display its grayscale
}
\examples{
grayscaleImage <- Rgrayscale("path_to_imageFile")

}
