% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rloadimage.R
\name{Rloadimage}
\alias{Rloadimage}
\title{Rloadimage}
\usage{
Rloadimage(imageFile)
}
\arguments{
\item{imageFile}{path to the image}
}
\description{
this should be internal function to load the image to use with other functions
}
\examples{
myObject <- Rloadimage("path_to_imageFile")

you need imagemagick installed, if you want to print the image, you need ggplot2
}
