% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{Rpkg.exam}
\alias{Rpkg.exam}
\title{Rpkg Exam Package}
\description{
Use this function
apply_filter("samples/Test image-1.jpg","grayscale")
apply_filter("samples/Test image-1.jpg","cutoff") #defaults to cutoff 127
apply_filter("samples/Test image-1.jpg","cutoff",20)
}
\details{
This package installs function Rgreyscale and Rsplit to be used by the user. See their description.
}
