% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsplit.R
\name{Rsplit}
\alias{Rsplit}
\title{Rsplit}
\usage{
Rsplit(imageFile)
}
\arguments{
\item{imageFile}{path to image to be converted to list of pictures (original picture and one per R G B channel)}
}
\value{
returns a list of images
}
\description{
this functions prints all three parts of RGB picture and returns them as list for futher work; you need imagemagick installed, if you want to print the image, you need ggplot2
}
\examples{
myList <- Rsplit("path_to_imageFile")

}
