% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE.R
\name{Rpkg.exam}
\alias{Rpkg.exam}
\title{Rpkg.exam}
\description{
This package serves the sole purpose of getting me through the course Intermediate R
On linux you might have to install libmagick++-dev
}
\examples{
Usage of this is package
  apply_filter("samples/Test image-1.jpg","grayscale")
  apply_filter("samples/Test image-1.jpg","cutoff") #defaults to cutoff 127
  apply_filter("samples/Test image-1.jpg","cutoff",20)

}
